Moved RegistrationSystem methods into OCServerData
