+++ server/src/main/java/com/omegaChess/server/OCServerData.java
@@ -11,7 +11,6 @@ class OCServerData {
   // fields
  // private final Logger log = LoggerFactory.getLogger(MicroServer.class);
   private ArrayList<UserProfile> profiles = new ArrayList<>();
-  private RegistrationSystem registrationSystem = new RegistrationSystem(this);
 
   // methods
 
@@ -19,8 +18,32 @@ class OCServerData {
     return profiles;
   }
 
-  public RegistrationSystem getRegistrationSystem() {
-    return registrationSystem;
+  public boolean createProfile(String nick, String pass, String email) {
+    if (!isNicknameTaken(nick)) {
+      UserProfile profile = new UserProfile(nick, pass, email);
+      getProfiles().add(profile);
+      return true;
+    }
+    return false;
+  }
+
+  private boolean isNicknameTaken(String nick) {
+    for (UserProfile profile : getProfiles()) {
+      if (profile.getNickname().equalsIgnoreCase(nick)) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  public boolean removeProfile(String nick) {
+    for (UserProfile profile : getProfiles()) {
+      if (profile.getNickname().equalsIgnoreCase(nick)) {
+        getProfiles().remove(profile);
+        return true;
+      }
+    }
+    return false;
   }
