--- server/src/main/java/com/omegaChess/server/RegistrationSystem.java
@@ -1,39 +1,0 @@
-package com.omegaChess.server;
-
-class RegistrationSystem {
-
-    private OCServerData server;
-
-    public RegistrationSystem(OCServerData serverReference) {
-        server = serverReference;
-    }
-
-    public boolean createProfile(String nick, String pass, String email) {
-        if (!isNicknameTaken(nick)) {
-            UserProfile profile = new UserProfile(nick, pass, email);
-            server.getProfiles().add(profile);
-            return true;
-        }
-        return false;
-    }
-
-    private boolean isNicknameTaken(String nick) {
-        for (UserProfile profile : server.getProfiles()) {
-            if (profile.getNickname().equalsIgnoreCase(nick)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    public boolean removeProfile(String nick) {
-        for (UserProfile profile : server.getProfiles()) {
-            if (profile.getNickname().equalsIgnoreCase(nick)) {
-                server.getProfiles().remove(profile);
-                return true;
-            }
-        }
-        return false;
-    }
-
-}
